+++ db_migrations/new.sql
@@ -1,0 +1,3 @@
+--add sprint tabs
+ALTER TABLE "project" ADD COLUMN "sprint_tabs" TEXT;
+--/add sprint tabs