+++ db_migrations/new.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "user" ADD COLUMN date_of_birth DATE DEFAULT NULL;