+++ db_migrations/new.sql
@@ -1,3 +1,4 @@
 --add sprint tabs
 ALTER TABLE "project" ADD COLUMN "sprint_tabs" TEXT;
 --/add sprint tabs
+ALTER TABLE "user" ADD COLUMN date_of_birth DATE DEFAULT NULL;