+++ src/intranet3/forms/user.py
@@ -23,6 +23,7 @@ class UserEditForm(wtf.Form):
     start_full_time_work  = wtf.DateField(_(u"Start full time work"), format='%d/%m/%Y', validators=[validators.Optional()])
     stop_work = wtf.DateField(_(u"Stop work"), format='%d/%m/%Y', validators=[validators.Optional()])
     description = wtf.TextField(_(u"Description"), validators=[validators.Optional()])
+    date_of_birth = wtf.DateField(_(u"Date of birth"), format='%d/%m/%Y', validators=[validators.Optional()])
     location = wtf.SelectField(
         _(u"Office location"),
         choices=[('', u'--None--')] + [(k, v[0]) for k, v in User.LOCATIONS.items()]