+++ CHANGES.md
@@ -1,3 +1,8 @@
+New release
+===========
+- default show active sprints only
+- project configuration allows to add custom tabs in sprint with embedded iframe
+
 0.9.5 (17-11-2013)
 ===============
 - much better scrum board layout